iomt-zero-trust/requirements.txt
fastapi
uvicorn[standard]
scapy
pyjwt[crypto]
pycryptodome
cryptography
wfdb
boto3
requests
flask
pytest
httpx
numpy
